index negative

+++ 0x17-doubly_linked_lists/5-get_dnodeint.c
@@ -13,7 +13,7 @@ dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 
 	if (head == NULL)
 		return (NULL);
-	if (index < 0)
+	if ((int)index < 0)
 		return (NULL);
 
 	c = head;
